Move panel related layout files to separate directory
