+++ components/draw/draw.module.ts
@@ -8,7 +8,7 @@ import {
   NgModule,
 } from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {HsPanelHelpersModule} from '../layout/panel-helpers.module';
+import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
 import {HsStylerModule} from '../styles/styles.module';
 
 import {FormsModule} from '@angular/forms';
